Revert "Disallow prohibited columns"
